refactoring controller tests, differentiating between Admin and User roles
