--- src/test/java/com/example/demo/controllers/AssociationControllerTest.java
@@ -1,28 +1,0 @@
-package com.example.demo.controllers;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class AssociationControllerTest {
-
-    @Test
-    void newAssociation() {
-    }
-
-    @Test
-    void updateAssociation() {
-    }
-
-    @Test
-    void deleteAssociation() {
-    }
-
-    @Test
-    void oneAssociation() {
-    }
-
-    @Test
-    void allAssociations() {
-    }
-}
