--- src/test/java/com/example/demo/controllers/MatchControllerTest.java
@@ -1,20 +1,0 @@
-package com.example.demo.controllers;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class MatchControllerTest {
-
-    @Test
-    void newMatch() {
-    }
-
-    @Test
-    void oneMatch() {
-    }
-
-    @Test
-    void allMatches() {
-    }
-}
