--- src/test/java/com/example/demo/controllers/SeasonControllerTest.java
@@ -1,20 +1,0 @@
-package com.example.demo.controllers;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class SeasonControllerTest {
-
-    @Test
-    void newSeason() {
-    }
-
-    @Test
-    void oneSeason() {
-    }
-
-    @Test
-    void allSeasons() {
-    }
-}
