+++ src/test/java/com/example/demo/controllers/userControllers/UserPlayerControllerTest.java
@@ -1,0 +1,24 @@
+package com.example.demo.controllers.userControllers;
+
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(webEnvironment= SpringBootTest.WebEnvironment.MOCK)
+@AutoConfigureMockMvc
+class UserPlayerControllerTest {
+
+    @Test
+    void onePlayer() {
+    }
+
+    @Test
+    void allPlayers() {
+    }
+}
